Some prep work for allowing login / logout, and authentication / authorization when it comes to adding new notes. There is now a login box on the base template and it seems to work fine.
Added a few fields on Video and Note in anticipation of future feature requests. There is now a flag to mark a video or a note "private" so that only the creator can see it. There is also now a field on Video that will theoretically lock notes from being added to it (except for the creator). None of the logic supporting these features is built yet, but the fields now exist.