Modified Note source_link to a char field rather than URL field to accomadate local paths (That maybe a django option somewhere, but whatever).
Modified Note save method to better populate offset, time, and link to note.

Will probably want to modify more to compensate for when 'time' isn't known on the video. Probably have to use save 'time' on video as the creation_time or something similar. Right now everything depends on it existing and it probably won't in all cases.

Added note saving to api resource. This was weird because the bundle wouldn't save properly on its own (it didn't bother to link the Video ID passed in to an actual video instance, so it was always null). Instead, manually creating a new Note object, applying the passed parameters, also pulling in the authenticated user and user_name and saving. Not sure what I should return here, so I'm just returning the new Note.

Added a note-taking view to the front end. Consists of a text box that saves when you hit enter. Notes get added back into the collection and when you close the note-taking view everything gets refreshed. This locks the UI up when there are a lot of notes, but I don't know a way around that (without ajaxing everything).

Added a (found) script that'll strip HTML tags from the note sent from the front end. It leaves a list of styling tags, but removes everything else. Hopefully.

CSS tweaks and other markup changes to make the note-taking view work properly.