Refactoring.
Basically moved the models / collections to a models.js file so I can include it in places where I'm not directly using the controllers (which would cause chaos to have them loaded). Also moved the Backbone sync override and a date util function to a utils.js file.

Moved some more generic Django->JS pieces from the video template to the base template since we'll be using that throughout the app (isLoggedIn, API urls).

Little bit of tweaking on the add video template, but that's not fleshed out at all at this point (refactoring the models came first).