Updated Resources to add User.
Basically included a subset of user properties on videos so we can figure out who exactly uploaded which video. Same for notes.

Note: Did not add these to User resource itself because if you include them every note a user ever entered is listed under every note. Need a way to make those properties only show up on a request straight for the user and not to something it's attached to. Maybe a separate resource entirely (Something like 'RelatedUserResource' for showing within relationships  and 'UserResource' for a details page  may work).

Removed some unnecessary CSS files.