Fix for user deletion on related saves
Basically, saving a video would fire off the 'save_related' method of the Resource. TastyPie does this blindly, even if you're manually specifiying 'fields' and then choosing 'full' for how to show the content. This means that anything not included in 'fields' will be saved as being blank. Basically it seems like they create a new instance of the model and manually assign the appropriate fields. When saved, it just overwrites whatever already exists. Yech.

For now I just totally wiped the save_related method from the Video and Note resources (it's just passed). We may have to change this someday (if we have models we actually want to save that way), but for now this works.

Modified add video js to not allow editing if video already exists.