obj_delete_list override. Styling
This is called by put_list, here: https://github.com/toastdriven/django-tastypie/blob/master/tastypie/resources.py#L1070

    I've run into an issue where all of the videos get deleted from the DB. this happens when Backbone "PUT"s a JSON request to the server
    without specifying an ID for the object. Basically, it's a bug on my part where Backbone thinks I'm updating but it's actually creating. BUT,
    it seems crazy that the entire video collection could be deleted so easily, and I don't want that to happen accidentally by anyone using the API.
    It also seems strange that this can be called even though 'delete' methods are specifically not allowed in the Meta.

    For now I'm going to not allow deleting an entire list to hopefully prevent this. Long-term would be to error-check in put_list as well as ensure
    Backbone doesn't put when I mean post (I have a fix for the Backbone part already). Also need to add user ownership verification.

==

Essentially TastyPie thinks I'm posting a list from BackBone (maybe I was, accidentally). It's default behavior on a "PUT" for a list is to delete the list (gah) and create a new one. I suspect that's the REST standard. BUT, I never want things to be deleted like that (and I specifically didn't include 'delete' in the resources, but that was denied).

So, removed 'put' permissions for lists. override obj_delete_list.

Some styling.