Source parsing. Removed force_insert in Note Save.
Note saving is done twice to create a convienience property for a link to the note within a video (you create a note while using the video interface, you want to have the link immediately for the video context, but there could also be a standalone note page). The first save gives you an ID so you can make the proper link. The problem is that when you use "get_or_create" the "force_insert" property is set to True on the kwargs. This means the first save inserts, and the second save ALSO insters, but with the ID property. This violates a key constraint and everything bombs.

Forced it to False in the Note save method. Hopefully this won't have any adverse effects.

Sources will now parse URLs on save. There's a lot of stuff that should be added to this. Right now it just checks that there's a url, a video, and a type. Then it tries to parse and sets scraped to True or False depending on how things go. We probably want to save some sort of failure message for the future, as well as run this entire process in some sort of queue so we don't delay page responses. Probably also want to make sure we're allowed to scrape the things we're trying to scrape.