Added note syncing to the video view.
Basically you click on a "Sync Notes" command and then click on a note to save the video with the new time. It figures out the offset of the video and the time of the note and figures out what time the video started at. Right now you have to manually refresh the page to see the notes synced up.

Also, had to change the server timezone away from Chicago because the times were all wrong and I almost broke the machine with fists of fury.