Removed "video length" math in video save method due to bug.
Basically, it was complaining that one date was 'offset-naive' and one was 'offset-aware' with regards to timestamps (as in, the 'time' had one and the 'end_time' didn't). Since we're not really using the end_time or the video_length at this point, I just commented it out rather than fixing it (Hate. Dates.).

Cleaned up the note syncing a bit, made the messaging and behavior a bit more friendly.