shortened the NOTE_SEEK_BEFORE_TIME to 7 seconds (from 30).
This is the time offset the video jumps to before the note that was clicked. It's hard to judge exactly what this time should be, and there's a good chance it varies from video to video and note-taker to note-taker. It may need to be customizable on a per-video basis, or have the option for a user to adjust it while they're working with the tool.