Added Video Updating to Video Page.
Basically if the logged in user owns the video being played, they can click on the title or description and change the details. This can stand in as a "video editing" page. The only editable fields are the title and description. Time is basically done by syncing notes, so manually typing it in isn't really necessary.

Moved jeditable from add_video template into base.

Fix bug on resources where I was testing for the "sync_notes" property incorrectly (using hasattr now).