Added a "note details" view.
Note details view shows sourcing (we need to make this consistent, right now it's a mess so it's hard to add actual hyperlinks or anything), photo if available. Could show any other details we wanted to collect. Also provides a method for "Syncing" the note to the current video position. This means you can move individual notes around in the timeline. Need to lock this down to the creating user / video owner.