Added note content editing.
Note details panel now edits-in-place like video details. Logic is if there is a user for the note (as in, they entered it through TVN) then they're the ones that can edit. If there is no user (as in, the note was imported from elsewhere) the video owner can edit it.