CSV Import, note syncing fix.
hasattr() apparently doesn't work the way I expected on the bundle.data. Changed to using get with a default of False for if we should sync notes or not.

Added an import function for CSVs in the same format as we export. This should let people make a backup copy of their notes.