Added Twitter parser. Error Message on sources.
Added an error_message property on Source so that if an exception is thrown during parsing at least we'll have some idea where to start. The exception text is just dumped into that field. Not sure if that's sane, but for now it works.

Fixed a stupid bug on Note for calculated offsets. It was getting the time of day right, but not the actual day (using the timedelta's seconds property and not the total_seconds. Oops.)

Twitter parser takes a Twitter user name, a start tweet and an end tweet. Optionally a hash tag. If there's no hash tag it'll take all of the tweets that come back (up to 200) and save them. Otherwise it'll search the hash element and pull in anything that matches.

Save method on Source removes any @ signs from user_name, #'s from hash, and can take a full Tweet link or just a Tweet ID (for usability reasons).