user.first_name + user.last_name takes precedence over user_name
In the video template we first check to see if a user exists (currently this would only happen if the user entered the note through the interface or the API -- Not Sources) and uses their first and last name as opposed to user_name. Not sure if this is the best logic ever, but it works for now.