Added User Page, Privacy Options Work
Added methods in views.py and resources.py that should always be used to get user visible objects before sending to the client. Basically they first decide to use the published objects manager or the normal manager (depending on your account status), and then will check that objects marked private are only shown if the associated user is the one that is currently logged in.

Added a User view. This just shows content from a specifc user.