Made NoteResource 'type' writeable.
I think the general idea here is that you should be able to give notes a "type" manually so you can filter on that type (as in, 'observation', or 'quote', or 'bill', things like that). As of now this is freeform, but maybe there should be a list.

Made sure the 'source' property gets saved on newly created notes.