Changed the get_user_visible_objects method.
Essentially made it so that the owner of the video can see all notes. The importer of a source can see all of the imported notes. I'm not sure if the video owner SHOULD see all notes, but for now this works. May be changed in the near future.

Added "Private" checkbox on note details display to toggle note privacy after-the-fact.