User added on Source Import
Made it so when a source is imported, that user is added to the imported objects (unless it is a CSV where the user id is included). If a source is not provided, the video user is used instead. This should mean that no Note is added to the system without a User (for the importers that have been updated. Storify needs a re-write and I'm not sure if we can even use Cover-It-Live).

Updated the video display template to use the user_name by default if it exists, and then the User property (this will make it so an imported source shows the original user, and not the user that created the object.