Made User required on Source
Made User non-nullable on the source object. Made it so if you import a CSV the user and user_name are automatically assigned to the Source creator, not out of the CSV (this is mostly to prevent people from assigning notes to others -- not sure of a better way to do that). Tested Scribblive parsing to make sure the user was coming from the right place.