Added "Name" field to Source, Added Source Syncing.
Name field had to be added because there was no other way to generate something a sane human could read. Added field to Admin, and video creation.

Added the option to sync an entire source to a video. This is so you can line up a specific source without syncing every note on the video. You can now sync either all notes, an individual note, or a note from a specific source.