Switched to using MediaElementPlayer for HTML5 Audio / Video.
This was because the VideoJS script doesn't support audio. The plugins that purported to support audio didn't work (they were ancient compared to the rest of the library). MediaElement seems to be fine for now. It also supports YouTube video, so we may want to get rid of the YouTube specific code and go solely with this (that option is poorly documented -- as in, not at all -- so I have to do some more investigation).