Added "Edit" functionality to Add page.
Basically rather than just not allowing editing on the add page if the video already exists, it instead checks to see if you're the owner of the video and displays it. It'll also display an Add Source box so another source can be added. I'm not sure if it is necessary yet to list all of the sources and allow them to be deleted (editing seems like a bad idea -- you'd have to delete existing notes and add new ones).

Added router to the add video page that'll examine the hash for #edit/VIDEO_ID and load the appropriate details.

Added link on video page to "edit video" for logged in owner of video.