Fixed the Storify parser to use the new API.
Also updated it to add proper data in newly added Note fields. Currently just grabs the max limit of elements Storify will respond with (1000), though pagination could be added in the future. Also only uses the Storify elements labelled as "quote." I have to find a list of them somewhere and see what else we could want, but this gets all of the Tweets...