Added Filters For UserName, Source, and Type.
When you hover over a note you'll see additional options to filter by that source, user, or type. The main router has had this added so those pages can be bookmarked.

Moved the actual filtering of the notes from the SearchView into the NotesView. Refactored to take a filter function and call that function against all notes. If it returns true, show it, if it returns false, hide it.

Added a 'reset' link above the search box.

On page load if no route is present the thing will automatically add a route for the 'all notes' view. Not sure if this is a good idea or not, but I wanted a way to back into the all notes view and this appeared to be the only way to do it.