Ajax uploader functionality.
Because for some reason the standard S3 upload backend from django-ajax-uploader didn't work properly, there are some workarounds (it got as far as uploading the files, but they always resulted in a size of 0 bytes). First the data is written to the /tmp/ folder. Then pushed to S3. Then deleted from the server. Then the JS gets the result notification, puts the proper URL into the source textbox and fires off the rest of the video creation. There is some logic to prevent files from being overwritten. Individual users uploads are kept in separate folders.