S3 Policy also includes username. Added datestamp to file upload.
So when the S3 policy for the uploader is generated it also includes the username in the key path. This way the user cannot fiddle with the settings to upload to another folder. A timestamp is also generated on the server for use in generating the filename (this is to prevent people from overriding their own files).