Virtual Scrolling. Sources not appearing on Edit Video page. Offset Weirdness on Source Sync.
The note list should now be virtualized. Essentially as the user scrolls the list calculates what is visible and renders those elements (as well as a few offscreen in either direction). This stops the horrible UI pauses from happening as the notes load or as people search when there are ~1000 items are more.

The current implementation isn't fantastic. The notes div is totally wiped out on every scroll and regenerated with the appropriate content. This is because I ran into all manner of weirdness trying to only remove the offscreen elements as the user scrolled (and adding new ones). Things like events not attaching, highlighted classes not working, etc. I'm sure some more effort spent on it would fix those bugs, but for now this works and the performance is actually pretty good.

-

Made sure that the Source window would appear when editing videos.

When syncing a source the note that was clicked would move to the correct time, but then would be moved back another amount of time based on whatever the offset was. This note is now excluded from the second time update (which moves all of the other Source notes into place). This should fix that issue.