Simple caption view.
It's hidden if there are no notes with the type "caption" -- this should mean that no one will see it until we provide a way to pull in captions. Currently we can do this via adding a Granicus Source in the backend. While it works, I'm not sure what we should do about the caption-notes in the normal note window. You want to be able to search them, but do you really want them to always be displayed? Maybe they should go away if you're doing anything other than a search?