Parsing for SRT caption files.
Added a requirement for pysrt, which converts an SRT caption file into Python objects. Added a parser that'll convert the SRT data into Notes. Add that to Source options.

Fixed a bug where editing a note in the details panel would not update the note in the Add Note pane or the notes list.

Added a migration adding a field to Source that should contain SRT pasted data. Added that to Admin as read-only (need to make a way to have TinyMCE not appear on that field and the csv_data field.

JS and templates for adding SRT data. CSS styling.