Added "import_source_name" to the Notes model, modified NoteResource to use fewer queries.
To generate the URL for an API call TastyPie needs to query the model and return the appropriate path. Unfortunately, this led to 4 queries for each note (1 each for Video, User, and Source, and a 4th to grab the import_source_name for me to display). This caused some serious slowness when you got over a few hundred notes on a page (at 1,000 you had to wait as long as 30seconds to a minute for the page to load). This wasn't really acceptable.

My solution was the use the _build_reverse_url method on the TastyPie resource model (example here: https://github.com/toastdriven/django-tastypie/issues/371) to avoid those queries. I then added an import_source_name property to the Note model to save the source name and avoid that extra query. That name was also added to all of the import scripts so it'll appear in future imports. For the meantime, the dehydrate method on the Notes Resource will add the name if it doesn't exist and save the note before delivering it (this is just a way to clean up existing notes).

JS modified to use a user_id property rather than look for a User model on each note. Also modified name of import_source_name (from source_title).

Migrations for import_source_name

import_source_name added to admin as read-only on Note.